# service-transaksi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaksi Parkir",
        "description": "API Transaksi untuk web DPark",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Admin memantau seluruh transaksi parkir",
                "operationId": "9115b43a55e6c92c81e6c62926e4034f",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Petugas mencatat kendaraan masuk melalui scan plat nomor",
                "operationId": "5968345d7524698be1b46f545cf2a41d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plate_number",
                                    "location_id",
                                    "vehicle_type"
                                ],
                                "properties": {
                                    "plate_number": {
                                        "type": "string",
                                        "example": "D 1234 ABC"
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_type": {
                                        "type": "string",
                                        "example": "mobil",
                                        "enum": [
                                            "motor",
                                            "mobil"
                                        ]
                                    },
                                    "member_id": {
                                        "type": "string",
                                        "example": "MBR-001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Petugas melihat detail transaksi dan total biaya parkir",
                "operationId": "688bc3b066f6fe867ee32162fa08e71b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Sistem memperbarui status transaksi setelah pembayaran berhasil",
                "operationId": "14886f2fd6159f736e882152b100b857",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "discount_amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "completed",
                                        "enum": [
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Input API Key here.",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Transactions",
            "description": "Transactions"
        }
    ]
}